+++ .eslintrc.js
@@ -4,6 +4,7 @@ module.exports = {
     node: true,
     "vue/setup-compiler-macros": true,
   },
+  ignorePatterns: ["dist/**/*"],
   extends: [
     "eslint:recommended",
     "plugin:vue/vue3-essential",
