+++ package.json
@@ -1,13 +1,12 @@
 {
   "name": "@xy-planning-network/trees",
-  "version": "0.3.4",
+  "version": "0.4.0-rc",
   "description": "",
   "license": "MIT",
   "repository": "github:xy-planning-network/trees",
   "files": [
     "config/*",
     "dist/*",
-    "trees.d.ts",
     "src/**/*.d.ts",
     "src/**/*.vue",
     "src/**/*.css"
@@ -16,14 +15,14 @@
   "module": "dist/trees.es.js",
   "style": "src/index.css",
   "unpkg": "dist/trees.umd.js",
-  "types": "trees.d.ts",
+  "types": "dist/types/entry.d.ts",
   "sideEffects": [
     "*.css",
     "*.vue"
   ],
   "scripts": {
     "dev": "vite",
-    "build": "vue-tsc --noEmit && vite build",
+    "build": "vite build && vue-tsc -d --emitDeclarationOnly && tsc-alias",
     "build:docs": "vue-tsc --noEmit && vite build --config vite.docs.config.ts",
     "lint": "eslint --ext .js,.ts,.vue src",
     "lint:fix": "eslint --fix --ext .js,.ts,.vue src dev && prettier -w -u src dev",
@@ -47,6 +46,7 @@
     "mitt": "^2.1.0",
     "postcss": "^8.4.5",
     "tailwindcss": "^2.2.19",
+    "tsc-alias": "^1.5.0",
     "typescript": "^4.4.4",
     "vite": "^2.7.2",
     "vite-plugin-environment": "^1.0.1",
