+++ src/types/nav.ts
@@ -1,0 +1,8 @@
+import { DefineComponent } from "vue"
+
+export interface Item {
+  icon?: DefineComponent<unknown, unknown, any>
+  name: string
+  openInTab?: boolean
+  url: string
+}
