--- src/types/users.d.ts
@@ -1,12 +1,0 @@
-declare namespace Users {
-  export interface User {
-    accountID: number
-    accountOwner: boolean
-    archived: boolean
-    id: number
-    email: string
-    name: string
-  }
-}
-
-export default Users
