+++ src/types/users.ts
@@ -1,0 +1,10 @@
+export interface User {
+  accountID: number
+  accountOwner: boolean
+  archived: boolean
+  id: number
+  email: string
+  name: string
+}
+
+export default User
